+++ src/renderer/Common/Icons/TabType/NewFile.svelte
@@ -1,0 +1,7 @@
+<script lang="ts">
+  let { size = "16" } = $props<{ size?: string }>();
+</script>
+
+<svg width={size} height={size} viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg">
+  <rect x="0" y="0" width="16" height="16" rx="4" fill="#5A5A5C" />
+</svg>